+++ include/mruby/presym.h
@@ -18,18 +18,21 @@
  * replaced by the symbol ID if presym is enabled by using the following
  * macros.
  *
- *   MRB_OPSYM(xor)  //=> ^      (Operator)
- *   MRB_CVSYM(xor)  //=> @@xor  (Class Variable)
- *   MRB_IVSYM(xor)  //=> @xor   (Instance Variable)
- *   MRB_SYM_B(xor)  //=> xor!   (Method with Bang)
- *   MRB_SYM_Q(xor)  //=> xor?   (Method with Question mark)
- *   MRB_SYM_E(xor)  //=> xor=   (Method with Equal)
- *   MRB_SYM(xor)    //=> xor    (Word characters)
+ *   MRB_OPSYM(xor)                //=> ^         (Operator)
+ *   MRB_SVSYM(argv)               //=> $*        (Special Variable)
+ *   MRB_CVSYM(xor)                //=> @@xor     (Class Variable)
+ *   MRB_IVSYM(xor)                //=> @xor      (Instance Variable)
+ *   MRB_GVSYM(xor), MRB_GVSYM(1)  //=> $xor, $1  (Global Variable)
+ *   MRB_SYM_B(xor)                //=> xor!      (Method with Bang)
+ *   MRB_SYM_Q(xor)                //=> xor?      (Method with Question mark)
+ *   MRB_SYM_E(xor)                //=> xor=      (Method with Equal)
+ *   MRB_SYM(xor)                  //=> xor       (Word characters)
  *
- * For `MRB_OPSYM`, specify the names corresponding to operators (see
- * `MRuby::Presym::OPERATORS` in `lib/mruby/presym.rb` for the names that
- * can be specified for it). Other than that, describe only word characters
- * excluding leading and ending punctuations.
+ * For `MRB_OPSYM` and `MRB_SVSYM`, specify the names corresponding to
+ * operators and special variables (see `MRuby::Presym::OPERATORS` and
+ * `MRuby::Presym::SPECIAL_VARIABLES` in `lib/mruby/presym.rb` for the names
+ * that can be specified for them). Other than that, describe only word
+ * characters excluding leading and ending punctuations.
  *
  * These macros are expanded to `mrb_intern_lit` if presym is disabled,
  * therefore the mruby state variable is required. The above macros can be