+++ include/mruby/presym/enable.h
@@ -10,16 +10,20 @@
 #include <mruby/presym/id.h>
 
 #define MRB_OPSYM(name) MRB_OPSYM__##name
+#define MRB_SVSYM(name) MRB_SVSYM__##name
 #define MRB_CVSYM(name) MRB_CVSYM__##name
 #define MRB_IVSYM(name) MRB_IVSYM__##name
+#define MRB_GVSYM(name) MRB_GVSYM__##name
 #define MRB_SYM_B(name) MRB_SYM_B__##name
 #define MRB_SYM_Q(name) MRB_SYM_Q__##name
 #define MRB_SYM_E(name) MRB_SYM_E__##name
 #define MRB_SYM(name) MRB_SYM__##name
 
 #define MRB_OPSYM_2(mrb, name) MRB_OPSYM__##name
+#define MRB_SVSYM_2(mrb, name) MRB_SVSYM__##name
 #define MRB_CVSYM_2(mrb, name) MRB_CVSYM__##name
 #define MRB_IVSYM_2(mrb, name) MRB_IVSYM__##name
+#define MRB_GVSYM_2(mrb, name) MRB_GVSYM__##name
 #define MRB_SYM_B_2(mrb, name) MRB_SYM_B__##name
 #define MRB_SYM_Q_2(mrb, name) MRB_SYM_Q__##name
 #define MRB_SYM_E_2(mrb, name) MRB_SYM_E__##name